Data.Map.Ordered

Description

An OMap behaves much like a Map, with mostly the same asymptotics, but also remembers the order that keys were inserted. All operations whose asymptotics are worse than Map have documentation saying so.

Documentation

data OMap k v

Trivial maps

empty :: OMap k v

singleton :: (k, v) -> OMap k v

Insertion

Conventions:

The open side of an angle bracket points to an OMap

The pipe appears on the side whose indices take precedence if both sides contain the same key

The left argument's indices are lower than the right argument's indices

If both sides contain the same key, the tuple's value wins

(<|) :: Ord k => (,) k v -> OMap k v -> OMap k v infixr 5

(|<) :: Ord k => (,) k v -> OMap k v -> OMap k v infixr 5

(>|) :: Ord k => OMap k v -> (,) k v -> OMap k v infixl 5

(|>) :: Ord k => OMap k v -> (,) k v -> OMap k v infixl 5

(<>|) :: Ord k => OMap k v -> OMap k v -> OMap k v infixr 6

When a key occurs in both maps, prefer the value from the first map.

See asymptotics of unionWithR.

(|<>) :: Ord k => OMap k v -> OMap k v -> OMap k v infixr 6

When a key occurs in both maps, prefer the value from the first map.

See asymptotics of unionWithL.

unionWithL :: Ord k => (k -> v -> v -> v) -> OMap k v -> OMap k v -> OMap k v

Take the union. The first OMap 's argument's indices are lower than the second. If a key appears in both maps, the first argument's index takes precedence, and the supplied function is used to combine the values.

O(r*log(r)) where r is the size of the result

unionWithR :: Ord k => (k -> v -> v -> v) -> OMap k v -> OMap k v -> OMap k v

Take the union. The first OMap 's argument's indices are lower than the second. If a key appears in both maps, the second argument's index takes precedence, and the supplied function is used to combine the values.

O(r*log(r)) where r is the size of the result

newtype Bias (dir :: IndexPreference) a

A newtype to hand a Monoid instance on. The phantom first parameter tells whether mappend will prefer the indices of its first or second argument if there are shared elements in both.

Constructors

Bias

Fields

unbiased :: a

type L = 'L

type R = 'R

Deletion

delete :: Ord k => k -> OMap k v -> OMap k v

filter :: Ord k => (k -> v -> Bool) -> OMap k v -> OMap k v

filter f m contains exactly the key-value pairs of m that satisfy f, without changing the order they appear

(\\) :: Ord k => OMap k v -> OMap k v' -> OMap k v

m \\ n deletes all the keys that exist in n from m

O(m*log(n)) where m is the size of the smaller map and n is the size of the larger map.

(|/\) :: Ord k => OMap k v -> OMap k v' -> OMap k v

Intersection. (The /\ is intended to look a bit like the standard mathematical notation for set intersection.)

See asymptotics of intersectionWith.

(/\|) :: Ord k => OMap k v -> OMap k v' -> OMap k v

Intersection. (The /\ is intended to look a bit like the standard mathematical notation for set intersection.)

See asymptotics of intersectionWith.

intersectionWith :: Ord k => (k -> v -> v' -> v'') -> OMap k v -> OMap k v' -> OMap k v''

Take the intersection. The first OMap 's argument's indices are used for the result.

O(m*log(n/(m+1)) + r*log(r)) where m is the size of the smaller map, n is the size of the larger map, and r is the size of the result.

Query

null :: OMap k v -> Bool

size :: OMap k v -> Int

member :: Ord k => k -> OMap k v -> Bool

notMember :: Ord k => k -> OMap k v -> Bool

lookup :: Ord k => k -> OMap k v -> Maybe v

Indexing

type Index = Int

A 0-based index, much like the indices used by lists' !! operation. All indices are with respect to insertion order.

findIndex :: Ord k => k -> OMap k v -> Maybe Index

elemAt :: OMap k v -> Index -> Maybe (k, v)

List conversions

fromList :: Ord k => [(k, v)] -> OMap k v

If a key appears multiple times, the first occurrence is used for ordering and the last occurrence is used for its value. The library author welcomes comments on whether this default is sane.

assocs :: OMap k v -> [(k, v)]

Return key-value pairs in the order they were inserted.

toAscList :: OMap k v -> [(k, v)]

Return key-value pairs in order of increasing key.

Map conversion

toMap :: OMap k v -> Map k v

Convert an OMap to a Map.

O(n), where n is the size of the OMap.