Data.Set.Ordered

Description

An OSet behaves much like a Set, with mostly the same asymptotics, but also remembers the order that values were inserted. All operations whose asymptotics are worse than Set have documentation saying so.

Documentation

data OSet a

Trivial sets

empty :: OSet a

singleton :: a -> OSet a

Insertion

Conventions:

The open side of an angle bracket points to an OSet

The pipe appears on the side whose indices take precedence for keys that appear on both sides

The left argument's indices are lower than the right argument's indices

(<|) :: Ord a => a -> OSet a -> OSet a infixr 5

(|<) :: Ord a => a -> OSet a -> OSet a infixr 5

(>|) :: Ord a => OSet a -> a -> OSet a infixl 5

(|>) :: Ord a => OSet a -> a -> OSet a infixl 5

(<>|) :: Ord a => OSet a -> OSet a -> OSet a infixr 6

O(m*log(n)+n), where m is the size of the smaller set and n is the size of the larger set.

(|<>) :: Ord a => OSet a -> OSet a -> OSet a infixr 6

O(m*log(n)+n), where m is the size of the smaller set and n is the size of the larger set.

newtype Bias (dir :: IndexPreference) a

A newtype to hand a Monoid instance on. The phantom first parameter tells whether mappend will prefer the indices of its first or second argument if there are shared elements in both.

Constructors

Bias

Fields

unbiased :: a

type L = 'L

type R = 'R

Query

null :: OSet a -> Bool

size :: OSet a -> Int

member :: Ord a => a -> OSet a -> Bool

notMember :: Ord a => a -> OSet a -> Bool

Deletion

delete :: Ord a => a -> OSet a -> OSet a

filter :: Ord a => (a -> Bool) -> OSet a -> OSet a

(\\) :: Ord a => OSet a -> OSet a -> OSet a

Set difference: r \\ s deletes all the values in s from r. The order of r is unchanged.

O(m*log(n)) where m is the size of the smaller set and n is the size of the larger set.

(|/\) :: Ord a => OSet a -> OSet a -> OSet a

Intersection. (/\ is meant to look a bit like the standard mathematical notation for intersection.)

O(m*log(n/(m+1)) + r*log(r)), where m is the size of the smaller set, n the size of the larger set, and r the size of the result.

(/\|) :: Ord a => OSet a -> OSet a -> OSet a

flip (|/\)

See asymptotics of |/\.

Indexing

type Index = Int

A 0-based index, much like the indices used by lists' !! operation. All indices are with respect to insertion order.

findIndex :: Ord a => a -> OSet a -> Maybe Index

elemAt :: OSet a -> Index -> Maybe a

List conversions

fromList :: Ord a => [a] -> OSet a

If a value occurs multiple times, only the first occurrence is used.

toAscList :: OSet a -> [a]

Returns values in ascending order. (Use toList to return them in insertion order.)

Set conversion

toSet :: OSet a -> Set a

Convert an OSet to a Set.

O(n), where n is the size of the OSet.